+++ R/cellxgene.R
@@ -164,6 +164,8 @@ ExtractCELLxGENEMeta <- function(all.samples.df, organism = NULL, ethnicity = NU
 #' @param quiet Logical value, whether to show downloading progress. Default: FALSE (show).
 #' @param parallel Logical value, whether to download parallelly. Default: TRUE. When "libcurl" is available for \code{download.file},
 #' the parallel is done by default (\code{parallel} can be FALSE).
+#' @param use.cores The number of cores used. Default: NULL (the minimum value of
+#' extracted \code{length(download.urls)} and \code{parallel::detectCores()}).
 #' @param return.seu Logical value, whether to load downloaded datasets to Seurat. Valid when rds in \code{file.ext} and all
 #' datasets download successfully. Default: FALSE.
 #' @param merge Logical value, whether to merge Seurat list when there are multiple rds files,
@@ -181,7 +183,7 @@ ExtractCELLxGENEMeta <- function(all.samples.df, organism = NULL, ethnicity = NU
 #' Default: FALSE (suppress loading in any dimensional reductions).
 #' @param ... Parameters for \code{\link{get_seurat}}, used when \code{use.census} is TRUE.
 #'
-#' @return Dataframe contains failed datasets, SeuratObject (\code{return.seu} is TRUE, rds in \code{file.ext}) and
+#' @return Dataframe contains failed datasets, SeuratObject (\code{return.seu} is TRUE, rds in \code{file.ext}) or
 #' NULL (\code{return.seu} is FALSE or rds not in \code{file.ext}).
 #' @importFrom httr POST stop_for_status content
 #' @importFrom jsonlite fromJSON
@@ -207,7 +209,7 @@ ExtractCELLxGENEMeta <- function(all.samples.df, organism = NULL, ethnicity = NU
 #' ParseCELLxGENE(meta = human.10x.cellxgene.meta, out.folder = "/path/to/output")
 #' }
 ParseCELLxGENE <- function(meta = NULL, file.ext = c("rds", "h5ad"), out.folder = NULL, timeout = 3600, quiet = FALSE,
-                           parallel = TRUE, return.seu = FALSE, merge = TRUE, use.census = FALSE, census.version = "stable",
+                           parallel = TRUE, use.cores = NULL, return.seu = FALSE, merge = TRUE, use.census = FALSE, census.version = "stable",
                            organism = NULL, obs.value.filter = NULL, obs.keys = NULL, include.genes = NULL, obsm.layers = FALSE, ...) {
   if (use.census) {
     message(
@@ -309,7 +311,7 @@ ParseCELLxGENE <- function(meta = NULL, file.ext = c("rds", "h5ad"), out.folder 
     message("Start downloading!")
     if (isTRUE(parallel)) {
       # prepare cores
-      cores.used <- min(parallel::detectCores(), length(download.urls))
+      cores.used <- min(parallel::detectCores(), length(download.urls), use.cores)
       down.status <- parallel::mclapply(X = 1:length(download.urls), FUN = function(x) {
         utils::download.file(url = download.urls[x], destfile = names(download.urls)[x], quiet = quiet, mode = "wb")
       }, mc.cores = cores.used)