+++ NEWS.md
@@ -7,6 +7,7 @@
 * Support `pigz` in `Bam2Fastq`.
 * Return full dataframe in `ExtractRun`.
 * Fix bug in `ParseCELLxGENE`, `ParseHCA`.
+* Add `use.cores` in `ParseCELLxGENE`, `ParseHCA`, `ParseZenodo`.
 
 -------------------